- Precalculate values for log(freq) -> screen_x mapping - Draw graph with optimizations - View model data is CGFloat (ready to display) - Model data is Float32 (internal type)

+++ FftMetalTest/SpectrumAnalyzer/Model/SpectrumAnalyzerDefaultInteractor.swift
@@ -8,5 +8,5 @@
 
 struct SpectrumAnalyzerDefaultInteractor: SpectrumAnalyzerInteractor {
     let isPlaying = Pipe<Bool>()
-    let samples = Pipe<[Double]>()
+    let samples = Pipe<[Float32]>()
 }
